refactor: 계산기 패키지로 클래스 옮김
