+++ src/main/java/study/calculator/domain/Addition.java
@@ -1,4 +1,4 @@
-package study.domain;
+package study.calculator.domain;
 
 import java.util.Arrays;
 import java.util.List;
