+++ src/test/java/study/AdditionTest.java
@@ -3,8 +3,8 @@ package study;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import study.domain.Addition;
-import study.domain.Operand;
+import study.calculator.domain.Addition;
+import study.calculator.domain.Operand;
 
 import static org.assertj.core.api.Assertions.assertThat;
