+++ src/test/java/study/OperandTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import study.domain.Operand;
+import study.calculator.domain.Operand;
 
 import static org.assertj.core.api.Assertions.*;
