+++ src/test/java/study/StringAdditionTest.java
@@ -3,7 +3,7 @@ package study;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import study.domain.StringAddition;
+import study.calculator.domain.StringAddition;
 
 import static org.assertj.core.api.Assertions.assertThat;
